+++ internal/coffee/controllers/controllers_recipe.go
@@ -1,4 +1,4 @@
-package coffee
+package coffee_controllers
 
 import (
 	"net/http"
