+++ internal/coffee/controllers/controllers_roaster.go
@@ -1,4 +1,4 @@
-package coffee
+package coffee_controllers
 
 import (
 	"fmt"
